feat: add mypy validation

+++ fias/importer/loader.py
@@ -4,57 +4,54 @@ from __future__ import unicode_literals, absolute_import
 import datetime
 import logging
 from sys import stdout
-from typing import List
+from typing import List, Any
 
 from django import db
 from django.conf import settings
-from django.core.exceptions import FieldDoesNotExist
 from django.db import IntegrityError
-from django.db.models import Model
 from progress import Infinite
 
-from fias.config import REMOVE_NOT_ACTUAL
-from fias.importer.signals import (
-    pre_import_table, post_import_table
-)
+from fias.config import REMOVE_NOT_ACTUAL, TableName
+from fias.importer.signals import pre_import_table, post_import_table
 from fias.importer.table.table import AbstractTableList, Table
-from fias.importer.validators import validators
-
+from fias.importer.validators import validate
+from fias.models import AbstractModel, AbstractIsActiveModel
 
 logger = logging.getLogger(__name__)
 
 
-class LoadingBar(Infinite):
+class LoadingBar(Infinite):  # type: ignore
     file = stdout
-    check_tty = False
+    check_tty: bool = False
 
-    text = 'T: %(table)s.' \
-           ' L: %(loaded)d | U: %(updated)d' \
-           ' | S: %(skipped)d[E:%(errors)d]' \
-           ' | R: %(depth)d[%(stack_str)s]' \
-           ' \tFN: %(filename)s'
+    text: str = 'T: %(table)s.' \
+                ' L: %(loaded)d | U: %(updated)d' \
+                ' | S: %(skipped)d[E:%(errors)d]' \
+                ' | R: %(depth)d[%(stack_str)s]' \
+                ' \tFN: %(filename)s'
 
-    loaded = 0
-    updated = 0
-    skipped = 0
-    errors = 0
-    depth = 0
-    stack = []
-    stack_str = 0
-    hide_cursor = False
+    loaded: int = 0
+    updated: int = 0
+    skipped: int = 0
+    errors: int = 0
+    depth: int = 0
+    stack: List[str]
+    stack_str: str = "0"
+    hide_cursor: bool = False
 
-    def __init__(self, message=None, **kwargs):
-        self.table = kwargs.pop('table', 'unknown')
-        self.filename = kwargs.pop('filename', 'unknown')
+    def __init__(self, message: str | None = None, table: str = 'unknown', filename: str = 'unknown', **kwargs: Any):
+        self.table = table
+        self.filename = filename
+        self.stack = []
         super(LoadingBar, self).__init__(message=message, **kwargs)
 
-    def __getitem__(self, key):
+    def __getitem__(self, key: str) -> Any:
         if key.startswith('_'):
             return None
         return getattr(self, key, None)
 
     def update(self, loaded: int = 0, updated: int = 0, skipped: int = 0, errors: int = 0, regress_depth: int = 0,
-               regress_len: int = 0, regress_iteration: int = 0):
+               regress_len: int = 0, regress_iteration: int = 0) -> None:
         if loaded:
             self.loaded = loaded
         if updated:
@@ -66,17 +63,17 @@ class LoadingBar(Infinite):
 
         self.depth = regress_depth
         if not self.depth:
-            self.stack_str = 0
+            self.stack_str = "0"
         else:
-            regress_len = '{0}:{1}'.format(regress_iteration, regress_len)
+            regress_len_s = f'{regress_iteration}:{regress_len}'
             stack_len = len(self.stack)
             if stack_len == self.depth:
-                self.stack[self.depth - 1] = regress_len
+                self.stack[self.depth - 1] = regress_len_s
             elif stack_len < self.depth:
-                self.stack.append(regress_len)
+                self.stack.append(regress_len_s)
             else:
                 self.stack = self.stack[0:self.depth]
-                self.stack[self.depth - 1] = regress_len
+                self.stack[self.depth - 1] = regress_len_s
 
             self.stack_str = '/'.join(self.stack)
 
@@ -94,13 +91,17 @@ class TableLoader(object):
         self.err_counter = 0
         self.today = datetime.date.today()
 
-    def validate(self, table: Table, item: Model):
-        if item is None or item.pk is None:
+    def validate(self, table: Table, item: AbstractModel) -> bool:
+        if item.pk is None:
             return False
 
-        return validators.get(table.name, lambda x, **kwargs: True)(item, today=self.today)
+        try:
+            tn = TableName(table.name)
+            return validate(tn, item, today=self.today)
+        except ValueError:
+            return True
 
-    def regressive_create(self, table: Table, objects: List, bar: LoadingBar, depth: int = 1):
+    def regressive_create(self, table: Table, objects: List[AbstractModel], bar: LoadingBar, depth: int = 1) -> None:
         count = len(objects)
         batch_len = count // 3 or 1
         batch_count = count // batch_len
@@ -122,7 +123,7 @@ class TableLoader(object):
                 else:
                     self.regressive_create(table, batch, bar=bar, depth=depth + 1)
 
-    def create(self, table: Table, objects: List, bar: LoadingBar):
+    def create(self, table: Table, objects: List[AbstractModel], bar: LoadingBar) -> None:
         try:
             table.model.objects.bulk_create(objects)
         except (IntegrityError, ValueError):
@@ -133,20 +134,20 @@ class TableLoader(object):
         if settings.DEBUG:
             db.reset_queries()
 
-    def load(self, tablelist: AbstractTableList, table: Table):
+    def load(self, tablelist: AbstractTableList, table: Table) -> None:
         logger.info(f'Region {table.region} table "{table.name}" is loading.')
         pre_import_table.send(sender=self.__class__, table=table)
         self.do_load(tablelist=tablelist, table=table)
         post_import_table.send(sender=self.__class__, table=table)
         logger.info(f'Region {table.region} table "{table.name}" has been loaded.')
 
-    def do_load(self, tablelist: AbstractTableList, table: Table):
+    def do_load(self, tablelist: AbstractTableList, table: Table) -> None:
         bar = LoadingBar(table=table.name, filename=table.filename)
         bar.update()
 
         objects = set()
         for item in table.rows(tablelist=tablelist):
-            if not self.validate(table, item):
+            if item is None or not self.validate(table, item):
                 self.skip_counter += 1
 
                 if self.skip_counter and self.skip_counter % self.limit == 0:
@@ -164,12 +165,8 @@ class TableLoader(object):
         if objects:
             self.create(table, list(objects), bar=bar)
 
-        if REMOVE_NOT_ACTUAL:
-            try:
-                _ = table.model._meta.get_field('isactive')
-                table.model.objects.filter(isactive=False).delete()
-            except FieldDoesNotExist:
-                pass
+        if REMOVE_NOT_ACTUAL and issubclass(table.model, AbstractIsActiveModel):
+            table.model.objects.filter(isactive=False).delete()
 
         bar.update(loaded=self.counter, skipped=self.skip_counter)
         bar.finish()
@@ -181,13 +178,13 @@ class TableUpdater(TableLoader):
         self.upd_limit = 100
         super(TableUpdater, self).__init__(limit=limit)
 
-    def do_load(self, tablelist: AbstractTableList, table: Table):
+    def do_load(self, tablelist: AbstractTableList, table: Table) -> None:
         bar = LoadingBar(table=table.name, filename=table.filename)
 
         model = table.model
         objects = set()
         for item in table.rows(tablelist=tablelist):
-            if not self.validate(table, item):
+            if item is None or not self.validate(table, item):
                 self.skip_counter += 1
                 continue
 
@@ -197,7 +194,7 @@ class TableUpdater(TableLoader):
                 objects.add(item)
                 self.counter += 1
             else:
-                if not hasattr(item, 'updatedate') or old_obj.updatedate < item.updatedate:
+                if old_obj.updatedate < item.updatedate:
                     item.save()
                     self.upd_counter += 1
 
@@ -212,12 +209,8 @@ class TableUpdater(TableLoader):
         if objects:
             self.create(table, list(objects), bar=bar)
 
-        if REMOVE_NOT_ACTUAL:
-            try:
-                _ = model._meta.get_field('isactive')
-                model.objects.filter(isactive=False).delete()
-            except FieldDoesNotExist:
-                pass
+        if REMOVE_NOT_ACTUAL and issubclass(model, AbstractIsActiveModel):
+            model.objects.filter(isactive=False).delete()
 
         bar.update(loaded=self.counter, updated=self.upd_counter, skipped=self.skip_counter)
         bar.finish()
